- Added side window - Added ability to change window colours
