+++ package.json
@@ -2,7 +2,7 @@
     "name": "openrct2-peep-editor",
     "author": "Manticore-007",
     "license": "MIT",
-    "version": "24.7.7",
+    "version": "24.10.27",
     "description": "",
     "main": "app.js",
     "scripts": {
