+++ src/helpers/progressBar.ts
@@ -1,85 +1,103 @@
 import {
-    Bindable,
-    Colour,
-    ElementParams,
-    FlexiblePosition,
-    WidgetCreator,
-    graphics,
-    isStore
-  } from "openrct2-flexui";
-  
-  interface ProgressBarParams extends ElementParams {
-    /**
-     * The background colour of the progress bar.
-     */
-    background: Bindable<Colour>
-  
-    /**
-     * The foreground colour of the progress bar.
-     */
-    foreground: Bindable<Colour>
-  
-    /**
-     * The percentage of the progress bar that is filled.
-     */
-    percentFilled: Bindable<number>
-  }
-  
-  function progressBar(
-    params: ProgressBarParams & FlexiblePosition
-  ): WidgetCreator<FlexiblePosition> {
-    return graphics({
-      width: params.width ?? "1w",
-      height: params.height ?? 10,
-      padding: {top: 2},
-      visibility: params.visibility || "visible",
-      disabled: params.disabled,
-      onDraw: (g) => {
-        const background: Colour = isStore(params.background)
-          ? params.background.get()
-          : params.background;
-        const foreground: Colour = isStore(params.foreground)
-          ? params.foreground.get()
-          : params.foreground;
-        const percentFilled = isStore(params.percentFilled)
-          ? params.percentFilled.get()
-          : params.percentFilled;
-        const disabled = isStore(params.disabled)
-          ? params.disabled.get()
-          : params.disabled;
-  
-        g.colour = background;
-        g.well(0, 0, 94, 10);
-        if (!disabled) {
-          g.colour = foreground;
-          g.box(1, 1, 94 * percentFilled - 2, 10 - 2);
-        } else {
-          g.box(1, 1, 94 - 2, 10 - 2);
-        }
-      }
-    });
-  }
-  
-  export { type ProgressBarParams, progressBar };
+  Bindable,
+  Colour,
+  ElementParams,
+  FlexiblePosition,
+  WidgetCreator,
+  graphics,
+  isStore,
+  store
+} from "openrct2-flexui";
+import { getColour } from "./settings";
 
-  // Define a function named percentage that calculates the percentage of a given number.
-export function percentage(fraction: number, total: number): number
-{
-  // Multiply num by per divided by 100 to find the percentage.
-  return fraction/total;
+
+export const ProgressBarColour = {
+  bar: {
+    safe: store<Colour>(getColour("pe.bar.safe", Colour.BrightGreen)),
+    warning: store<Colour>(getColour("pe.bar.warning", Colour.Yellow)),
+    danger: store<Colour>(getColour("pe.bar.danger", Colour.BrightRed)),
+  },
+  foreground: store<Colour>(Colour.DarkYellow),
+  background: store<Colour>(getColour("pe.bar.background", Colour.DarkYellow)),
 }
 
-export function colourPogressBar(positive: boolean, percentage: number): number
-{
-    if (positive){
-        if (percentage <= 0.5) return Colour.BrightRed;
-        else if (percentage > 0.5 && percentage < 0.7) return Colour.Yellow;
-        else if (percentage >= 0.7) return Colour.BrightGreen;
+
+interface ProgressBarParams extends ElementParams {
+  /**
+   * The background colour of the progress bar.
+   */
+  background: Bindable<Colour>
+
+  /**
+   * The foreground colour of the progress bar.
+   */
+  foreground: Bindable<Colour>
+
+  /**
+   * The percentage of the progress bar that is filled.
+   */
+  percentFilled: Bindable<number>
+
+  /**
+   * When true, full bar is positive.
+   * When false empty bar is positive.
+   */
+  isPositive: Bindable<boolean>
+}
+
+function progressBar(
+  params: ProgressBarParams & FlexiblePosition
+): WidgetCreator<FlexiblePosition> {
+  return graphics({
+    width: params.width ?? "1w",
+    height: params.height ?? 12,
+    padding: { top: 2 },
+    visibility: params.visibility || "visible",
+    disabled: params.disabled,
+    onDraw: (g) => {
+      const background: Colour = isStore(params.background)
+        ? params.background.get()
+        : params.background;
+      const percentFilled = isStore(params.percentFilled)
+        ? params.percentFilled.get()
+        : params.percentFilled;
+      const disabled = isStore(params.disabled)
+        ? params.disabled.get()
+        : params.disabled;
+        const isPositive = isStore(params.isPositive)
+          ? params.isPositive.get()
+          : params.isPositive;
+
+      g.colour = background;
+      g.well(0, 0, 88, 11);
+      if (!disabled) {
+        g.colour = colourPogressBar(isPositive, percentFilled);
+        g.box(1, 1, 88 * percentFilled, 11 - 2);
+      } else {
+        g.box(1, 1, 88, 11 - 2);
+      }
     }
-    else {
-        if (percentage <= 0.5) return Colour.BrightGreen;
-        else if (percentage > 0.5 && percentage < 0.7) return Colour.Yellow;
-        else if (percentage >= 0.7) return Colour.BrightRed;
-    }
-    return Colour.DarkYellow;
+  });
+}
+
+export { type ProgressBarParams, progressBar };
+
+// Define a function named percentage that calculates the percentage of a given number.
+export function percentage(fraction: number, total: number): number {
+  // Multiply num by per divided by 100 to find the percentage.
+  return fraction / total;
+}
+
+function colourPogressBar(positive: boolean, percentage: number): number {
+  if (positive) {
+    if (percentage <= 0.5) return ProgressBarColour.bar.danger.get();
+    else if (percentage > 0.5 && percentage < 0.7) return ProgressBarColour.bar.warning.get();
+    else if (percentage >= 0.7) return ProgressBarColour.bar.safe.get();
+  }
+  else {
+    if (percentage <= 0.5) return ProgressBarColour.bar.safe.get();
+    else if (percentage > 0.5 && percentage < 0.7) return ProgressBarColour.bar.warning.get();
+    else if (percentage >= 0.7) return ProgressBarColour.bar.danger.get();
+  }
+  return Colour.DarkYellow;
 }
