+++ src/main.ts
@@ -1,8 +1,10 @@
+import { Colour } from "openrct2-flexui";
 import { initActions } from "./actions/initActions";
 import { initCustomSprites } from "./helpers/customImages";
 import { isUiAvailable } from "./helpers/environment";
 import { debug } from "./helpers/logger";
-import { windowPeepEditor } from "./ui/mainWindow";
+import { mainWindow } from "./ui/mainWindow";
+import { getColour } from "./helpers/settings";
 
 /**
  * Entry point of the plugin.
@@ -17,5 +19,9 @@ export function main(): void
 	}
 	initActions();
 	initCustomSprites();
-	ui.registerMenuItem("Peep Editor", () => windowPeepEditor.open());
+	getColour("pe.main.primary", Colour.DarkYellow);
+	getColour("pe.main.secondary", Colour.DarkYellow);
+	getColour("pe.side.primary", Colour.DarkYellow);
+	getColour("pe.side.secondary", Colour.DarkYellow);
+	ui.registerMenuItem("Peep Editor (new UI)", () => mainWindow.open());
 }
